rearrange the journey components
remove comma
